Add files via upload
This is a "skeleton" of an program architecture discussed in class.  It has 3 DLLs and one executable.  Each team should be able to look at the CrossLevel.cs file to figure out how to put a class into a DLL.  (Externally, the classes are accessible with by "public static" access.)